JE SUIS DEBILE

+++ includes/cub3d_bonus.h
@@ -6,7 +6,7 @@
 /*   By: pkorsako <pkorsako@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/20 17:15:52 by pkorsako          #+#    #+#             */
-/*   Updated: 2024/03/28 16:22:42 by ajeannin         ###   ########.fr       */
+/*   Updated: 2024/04/03 18:38:55 by ajeannin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -123,8 +123,8 @@ typedef struct s_game
 	char		floor[10];
 	char		ceiling[10];
 	char		wall[10];
-	int			floor_x;
-	int			floor_y;
+	double			floor_x;
+	double			floor_y;
 	t_tex		*tex;
 	t_player	*player;
 	t_grid		*grid;
